% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sheet_data.R
\name{extract_sheet_data}
\alias{extract_sheet_data}
\title{Extract blocks of data from xlsx file exported from Belysa}
\usage{
extract_sheet_data(file, sheet, numeric_data = TRUE)
}
\arguments{
\item{file}{Character string with path to excel file to be read.}

\item{sheet}{Character string of the sheet to be read.}

\item{numeric_data}{Whether the data being read are numeric (default = \code{TRUE}).}
}
\value{
A tibble of sheet data data.
}
\description{
this is a description
}
\examples{
1 + 1
}
