% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_well_data.R
\name{extract_well_data}
\alias{extract_well_data}
\title{Title}
\usage{
extract_well_data(..., numeric_data = TRUE)
}
\arguments{
\item{...}{blah blah blah}

\item{numeric_data}{blah blah blah}
}
\value{
A tibble of well data
}
\description{
Title
}
\examples{
1 + 1
}
