% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytes.R
\name{get_analytes}
\alias{get_analytes}
\title{Extract the table of analytes from intelliframe}
\usage{
get_analytes(.intelliframe)
}
\arguments{
\item{.intelliframe}{An intelliframe object.}
}
\value{
A tibble
}
\description{
Convenience function to extract the table of analytes from an intelliframe object
}
\examples{
1+1
}
