% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curve_data.R
\name{get_curve_data}
\alias{get_curve_data}
\title{Extract the standard curve data from intelliframe}
\usage{
get_curve_data(.data)
}
\arguments{
\item{.data}{}
}
\value{
A tibble
}
\description{
Convenience function to extract the standard curve data from an intelliframe object
}
\examples{
1+1
}
