% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expected.R
\name{get_expected}
\alias{get_expected}
\title{Extract the table of expected concentrations from intelliframe}
\usage{
get_expected(.data)
}
\arguments{
\item{.data}{}
}
\value{
A tibble
}
\description{
Convenience function to extract the table of expected standard concentrations from an intelliframe object
}
\examples{
1+1
}
