% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract instrument metadata from intelliframe}
\usage{
get_metadata(.data)
}
\arguments{
\item{.data}{}
}
\value{
A tibble
}
\description{
Convenience function to extract instrument metadata from an intelliframe object
}
\examples{
1+1
}
