% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery.R
\name{get_recovery}
\alias{get_recovery}
\title{Extract the table of standard recoveries from intelliframe}
\usage{
get_recovery(.data)
}
\arguments{
\item{.data}{}
}
\value{
A tibble
}
\description{
Convenience function to extract the table of standard recoveries from an intelliframe object
}
\examples{
1+1
}
