% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_avg.R
\name{get_recovery_avg}
\alias{get_recovery_avg}
\title{Extract the table of average standard recoveries from intelliframe}
\usage{
get_recovery_avg(.data)
}
\arguments{
\item{.data}{}
}
\value{
A tibble
}
\description{
Convenience function to extract the table of average standard recoveries from an intelliframe object
}
\examples{
1+1
}
