% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_data.R
\name{get_summary_data}
\alias{get_summary_data}
\title{Extract summary data from intelliframe object}
\usage{
get_summary_data(.intelliframe)
}
\arguments{
\item{.intelliframe}{An intelliframe object.}
}
\value{
A tibble
}
\description{
Convenience function to extract summary data from an intelliframe object
}
\examples{
1+1
}
