% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_well_data.R
\name{get_well_data}
\alias{get_well_data}
\title{Extract individual well data from intelliframe object}
\usage{
get_well_data(.intelliframe)
}
\arguments{
\item{.intelliframe}{An intelliframe object.}
}
\value{
A tibble
}
\description{
Convenience function to extract individual well data from an intelliframe object
}
\examples{
1+1
}
