% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intelliframe.R
\name{intelliframe-class}
\alias{intelliframe-class}
\alias{intelliframe}
\title{intelliframe class}
\usage{
intelliframe(
  metadata = class_missing,
  analytes = class_missing,
  expected = class_missing,
  recovery = class_missing,
  recovery_avg = class_missing,
  well_data = class_missing,
  summary_data = class_missing,
  curve_data = class_missing
)
}
\description{
Defines the constructor function for the intelliframe S7 class. Not meant to be
called directly by the user (see \code{\link{read_xmap}}).
}
