% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intelliframe.R
\name{intelliframe}
\alias{intelliframe}
\title{intelliframe class}
\usage{
intelliframe(
  metadata = class_missing,
  analytes = class_missing,
  expected = class_missing,
  recovery = class_missing,
  recovery_avg = class_missing,
  well_data = class_missing,
  summary_data = class_missing,
  curve_data = class_missing
)
}
\arguments{
\item{metadata}{blah blah}

\item{analytes}{blah blah}

\item{expected}{blah blah}

\item{recovery}{blah blah}

\item{recovery_avg}{blah blah}

\item{well_data}{blah blah}

\item{summary_data}{blah blah}

\item{curve_data}{blah blah}
}
\description{
Defines the constructor function for the intelliframe S7 class. Not meant to be
called directly by the user (see \code{\link{read_xmap}}).
}
