% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves.R
\name{plot_curves}
\alias{plot_curves}
\title{Plot standard curves with interpolated values}
\usage{
plot_curves(
  .data,
  type = "individual",
  interactive = FALSE,
  point_size = 3,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
ggplot object
}
\description{
Plot standard curves with interpolated values
}
\examples{
1 + 1
}
