% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.intelliframe}
\alias{print.intelliframe}
\title{print method for intelliframe class}
\arguments{
\item{intelliframe}{}
}
\value{
printed summary
}
\description{
print method for intelliframe class
}
\examples{
1+1
}
