% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{print method for intelliframe class}
\arguments{
\item{intelliframe}{An object of class \code{intelliframe}, usually created using \code{\link{read_xmap}}.}
}
\value{
printed summary
}
\description{
print method for intelliframe class
}
\examples{
1+1
}
