% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.intelliframe}
\alias{str.intelliframe}
\title{str method for intelliframe class}
\usage{
\method{str}{intelliframe}(object, nest.lev = 0, ...)
}
\arguments{
\item{object}{An intelliframe object.}

\item{nest.lev}{How many levels to print.}

\item{...}{Unused.}
}
\description{
str method for intelliframe class
}
