% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_curve_data.R
\name{update_curve_data}
\alias{update_curve_data}
\title{update_curve_data}
\usage{
update_curve_data(.summary_data, .well_data, .fits, .silent)
}
\arguments{
\item{.summary_data}{A tibble containing the summary_data property.}

\item{.well_data}{A tibble containing the well_data property.}

\item{.fits}{A list of model fits returned by \code{nplr::nplr()}.}

\item{.silent}{Should luminary warnings be printed?}
}
\value{
A tibble
}
\description{
Update the curve_data property from an intelliframe object using a list of
model fits. Not meant to be called by the user.
}
