% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_recovery.R
\name{update_recovery}
\alias{update_recovery}
\title{update_recovery}
\usage{
update_recovery(.well_data)
}
\arguments{
\item{.well_data}{A tibble containing the well_data property.}
}
\value{
A tibble
}
\description{
Update the recovery property from an intelliframe object using up-to-date
well_data. Not meant to be called by the user.
}
