% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_recovery_avg.R
\name{update_recovery_avg}
\alias{update_recovery_avg}
\title{update_recovery_avg}
\usage{
update_recovery_avg(.recovery, .recovery_avg, .use_excluded, .excluded_wells)
}
\arguments{
\item{.recovery}{A tibble containing the original recovery_avg property.}

\item{.use_excluded}{Logical flag indicating whether wells with a value of
\code{TRUE} in the \code{Excluded} column are used to calculate summary
statistics.}

\item{.excluded_wells}{Logical vector indicating whether each well is
excluded.}
}
\value{
A tibble
}
\description{
Update the recovery_avg property from an intelliframe object using up-to-date
recovery. Not meant to be called by the user.
}
