% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summary_data.R
\name{update_summary_data}
\alias{update_summary_data}
\title{update_summary_data}
\usage{
update_summary_data(.well_data, .summary_data, .use_excluded, .excluded_wells)
}
\arguments{
\item{.well_data}{A tibble containing the well_data property.}

\item{.summary_data}{A tibble containing the original summary_data property.}

\item{.use_excluded}{Logical flag indicating whether wells with a value of
\code{TRUE} in the \code{Excluded} column are used to calculate summary
statistics.}

\item{.excluded_wells}{Logical vector indicating whether each well is
excluded.}
}
\value{
A tibble
}
\description{
Update the summary_data property from an intelliframe object using up-to-date
well_data Not meant to be called by the user.
}
