% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_well_data.R
\name{update_well_data}
\alias{update_well_data}
\title{update_well_data}
\usage{
update_well_data(.well_data, .fits, .standard_list, .silent)
}
\arguments{
\item{.well_data}{A tibble containing the well_data property.}

\item{.fits}{A list of model fits.}

\item{.standard_list}{A tibble of standard curve data.}

\item{.silent}{Logical flag indicating whether warnings and messages during
interpolation should be silenced.}
}
\value{
A tibble
}
\description{
Update the well_data property from an intelliframe object using a list of
model fits. Not meant to be called by the user.
}
