% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sheet_data.R
\name{extract_sheet_data}
\alias{extract_sheet_data}
\title{something something}
\usage{
extract_sheet_data(file, sheet, numeric_data = TRUE)
}
\arguments{
\item{numeric_data}{blah blah blah}

\item{...}{blah blah blah}
}
\value{
A tibble of well data
}
\description{
something something
}
\examples{
1 + 1
}
