% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sheet_data.R
\name{extract_sheet_data}
\alias{extract_sheet_data}
\title{Extract blocks of data from an xlsx file exported from Belysa}
\usage{
extract_sheet_data(file, sheet, numeric_data = TRUE)
}
\arguments{
\item{file}{Character string with path to excel file to be read.}

\item{sheet}{Character string of the sheet to be read.}

\item{numeric_data}{Whether the data being read are numeric (default = \code{TRUE}).}
}
\value{
A tibble.
}
\description{
Read an single sheet from an .xlsx file exported from the Belysa software,
and return an intelliframe object. Not meant to be called by the user.
}
\examples{
1 + 1
}
