% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intelliframe.R
\name{intelliframe}
\alias{intelliframe}
\title{intelliframe class}
\usage{
intelliframe(
  metadata = class_missing,
  summary = class_missing,
  expected = class_missing,
  mfi = class_missing,
  mfi_avg = class_missing,
  mfi_cv = class_missing,
  mfi_sd = class_missing,
  result = class_missing,
  result_avg = class_missing,
  result_cv = class_missing,
  result_sd = class_missing,
  recovery = class_missing,
  recovery_avg = class_missing,
  messages = class_missing,
  curve_data = class_missing,
  analytes = class_missing,
  excluded_wells = class_missing
)
}
\description{
Defines the constructor function for the intelliframe S7 class. Not meant to be
called directly by the user (see \code{\link{read_xmap}}).
}
\section{Slots}{

\describe{
\item{\code{metadata}}{is this how I describe each of the properties?}
}}

