% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xmap.R
\name{read_xmap}
\alias{read_xmap}
\title{Read data from xMAP instrument}
\usage{
read_xmap(file)
}
\arguments{
\item{file}{Path to the file to be read as a character string.}
}
\value{
A tibble
}
\description{
Read data from xMAP instrument (this is the description)
}
\examples{
1 + 1
}
