% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_curves.R
\name{refit_curves}
\alias{refit_curves}
\title{Refit standard curves}
\usage{
refit_curves(
  .intelliframe,
  npars = "all",
  weight_method = "res",
  LPweight = 0.25,
  add_to_zeroes = 0.02,
  silent = FALSE
)
}
\arguments{
\item{.intelliframe}{Intelliframe object to refit}

\item{npars}{A numeric value (or \code{"all"}) to specify the number of
parameters to use in the model. If \code{"all"} the logistic model will be
tested with 2 to 5 parameters, and the best option will be returned.
See Details.}

\item{weight_method}{A character string to specify what weight method to use.
Options are \code{"res"}(default), \code{"sdw"}, \code{"gw"}. See Details.}

\item{LPweight}{a coefficient to adjust the weights. \code{LPweight = 0} will
compute a non-weighted np-logistic regression.}

\item{add_to_zeroes}{A numeric value (\code{0.1} by default) added to zeros during
the curve fitting procedure as it requires the concentration to be
log10-transformed.}

\item{silent}{Logical flag indicating whether warnings and/or messages during
curve fitting should be silenced. Defaults to \code{FALSE}.}
}
\value{
An intelliframe
}
\description{
\loadmathjax
Function to refit n-parameter logistic curves to standards in an intelliframe object
and interpolate concentrations for all samples, based on the updated models.
}
\details{
Most of these details have been reproduced from the \code{\link{nplr}} function that is doing most of the heavy lifting.

The 5-parameter logistic regression is of the form:
\mjdeqn{y = B + (T - B)/[1 + 10^{(b*(x_{\text{mid}} - x))}]^s}{}

where \mjeqn{B}{} and \mjeqn{T}{} are the bottom and top asymptotes, respectively,
\mjeqn{b}{} and \mjeqn{x_{\text{mid}}}{} are the Hill slope and the x-coordinate at
the inflection point, respectively, and \mjeqn{s}{} is an asymmetric coefficient.
This equation is sometimes referred to as the Richards' equation [1,2].
When specifying \code{npars = 4}, the \mjeqn{s}{} parameter is forced to be 1,
and the corresponding model is a 4-parameter logistic regression, symmetrical
around its inflection point. When specifying \code{npars = 3} or \code{npars = 2},
add 2 more constraints and force \mjeqn{B}{} and \mjeqn{T}{} to be 0 and 1, respectively.

Weight methods:

The model parameters are optimized, simultaneously, using \code{\link{nlm}},
given a sum of squared errors function, \mjeqn{\text{sse(Y)}}{}, to minimize:
\mjdeqn{\text{sse}(Y) = \sum{W(Y_{\text{obs}} - Y_{\text{fit}})^2 }}{}
where \mjeqn{Y_{\text{obs}}}{}, \mjeqn{Y_{\text{fit}}}{} and \mjeqn{W}{} are the vectors
of observed values, fitted values and weights, respectively.
In order to reduce the effect of possible outliers, the weights can be computed in different ways:

residual weights, \code{"res"}:
\mjdeqn{W = (1/\text{residuals})^\text{LPweight}}{}
where \mjeqn{\text{residuals}}{} and \mjeqn{\text{LPweight}}{} are the squared error
between the observed and fitted values, and a tuning parameter, respectively.
Best results are generally obtained by setting \code{LPweight = 0.25} (default value),
while setting \code{LPweight = 0} results in computing a non-weighted sum of squared errors.

standard weights, \code{"sdw"}:
\mjdeqn{W = 1/\text{Var}(Y_{\text{obs}_r})}{}
where \mjeqn{\text{Var}(Y_{\text{obs}_r})}{} is the vector of the within-replicates variances.

general weights, \code{"gw"}:
\mjdeqn{W = 1/Y_{\text{fit}}^\text{LPweight}}{}
where \mjeqn{Y_{\text{fit}}}{} are the fitted values. As for the residuals-weights method,
setting \mjeqn{\text{LPweight} = 0}{} results in computing a non-weighted sum of squared errors.
The standard weights and general weights methods are described in [3].
}
\examples{
1+1
}
\references{
1- Richards, F. J. (1959). A flexible growth function for empirical use. J Exp Bot 10, 290-300.

2- Giraldo J, Vivas NM, Vila E, Badia A. Assessing the (a)symmetry of
concentration-effect curves: empirical versus mechanistic models. Pharmacol Ther. 2002 Jul;95(1):21-45.

3- Motulsky HJ, Brown RE. Detecting outliers when fitting data with nonlinear
regression - a new method based on robust nonlinear regression and the false
discovery rate. BMC Bioinformatics. 2006 Mar 9;7:123.
}
